Fix TypeError in OnPaint - convert column width to int
- Fixed: SetColumnWidth() argument must be int, not float
- Changed: width/4 → int(width/4) in OnPaint method
- Resolves: TypeError when resizing window or painting list

This was causing repeated errors in the console during normal operation.
Python 3 division returns float, but wxPython requires int for column width.

+++ wxAsyncNewsReader.py
@@ -340,7 +340,7 @@ class NewsPanel(wx.Panel):
     def OnPaint(self, evt):
         width, height = self.news_list.GetSize()
         for i in range(4):
-            self.news_list.SetColumnWidth(i, width/4)
+            self.news_list.SetColumnWidth(i, int(width/4))
         evt.Skip()
     
     def OnSourceSelected(self, event):
